[feat] 악보 선택 모달 모듈화
- constant 변수 모음
- localStorage에 업로드한 악보 관리하는 기능 모음
- 이벤트 핸들러
- 악보 선택 카드 렌더
