+++ src/components/common/music-sheet-select/constants.js
@@ -1,0 +1,34 @@
+// localStorage에 저장할 키 값
+export const LOCAL_STORAGE_KEY = "userMusicSheets";
+
+// 기본 제공 악보 데이터
+export const DEFAULT_MUSIC_SHEETS = [
+  {
+    id: "default-1",
+    name: "airplane",
+    src: "/assets/images/airplane.png",
+    type: "image",
+    isDeletable: false,
+  },
+  {
+    id: "default-2",
+    name: "butterfly",
+    src: "/assets/images/butterfly.png",
+    type: "image",
+    isDeletable: false,
+  },
+  {
+    id: "default-3",
+    name: "little-star",
+    src: "/assets/pdf/little-star.pdf",
+    type: "pdf",
+    isDeletable: false,
+  },
+  {
+    id: "default-4",
+    name: "three-bears",
+    src: "/assets/images/three-bears.png",
+    type: "image",
+    isDeletable: false,
+  },
+];
